Replace index with new
