--- pages/new.tsx
@@ -1,426 +1,0 @@
-import { Menu, Popover, Transition } from "@headlessui/react";
-import {
-  BellIcon,
-  FireIcon,
-  HomeIcon,
-  MenuIcon,
-  TrendingUpIcon,
-  UserGroupIcon,
-  XIcon,
-} from "@heroicons/react/outline";
-import { SearchIcon } from "@heroicons/react/solid";
-import classnames from "classnames";
-import { Fragment } from "react";
-import { TabPanel, useTabs } from "react-headless-tabs";
-import { AboutMe } from "../components/AboutMe";
-import { Experience } from "../components/Experience";
-import { Projects } from "../components/Projects";
-import { TechnicalSkills } from "../components/TechnicalSkills";
-import { Timeline } from "../components/Timeline";
-
-const user = {
-  name: "Chelsea Hagon",
-  email: "chelsea.hagon@example.com",
-  imageUrl:
-    "https://images.unsplash.com/photo-1550525811-e5869dd03032?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
-};
-
-const navigation = [
-  { name: "Home", href: "#", icon: HomeIcon, current: true },
-  { name: "Popular", href: "#", icon: FireIcon, current: false },
-  { name: "Communities", href: "#", icon: UserGroupIcon, current: false },
-  { name: "Trending", href: "#", icon: TrendingUpIcon, current: false },
-];
-
-const userNavigation = [
-  { name: "Your Profile", href: "#" },
-  { name: "Settings", href: "#" },
-  { name: "Sign out", href: "#" },
-];
-
-const communities = [
-  { name: "Movies", href: "#" },
-  { name: "Food", href: "#" },
-  { name: "Sports", href: "#" },
-  { name: "Animals", href: "#" },
-  { name: "Science", href: "#" },
-  { name: "Dinosaurs", href: "#" },
-  { name: "Talents", href: "#" },
-  { name: "Gaming", href: "#" },
-];
-
-const tabs = [
-  { name: "About me", href: "#about" },
-  { name: "Projects", href: "#projects" },
-  { name: "Experience", href: "#experience" },
-];
-
-export default function New() {
-  const [tab, setTab] = useTabs(
-    tabs.map(({ href }) => href),
-    typeof window === "undefined" ? undefined : window.location.hash
-  );
-
-  return (
-    <>
-      {/*
-        This example requires updating your template:
-
-        ```
-        <html class="h-full bg-gray-100">
-        <body class="h-full">
-        ```
-      */}
-      <div className="min-h-full">
-        {/* When the mobile menu is open, add `overflow-hidden` to the `body` element to prevent double scrollbars */}
-        <Popover
-          as="header"
-          className={({ open }) =>
-            classnames(
-              open ? "fixed inset-0 z-40 overflow-y-auto" : "",
-              "bg-white shadow-sm lg:static lg:overflow-y-visible"
-            )
-          }
-        >
-          {({ open }) => (
-            <>
-              <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-                <div className="relative flex justify-between xl:grid xl:grid-cols-12 lg:gap-8">
-                  <div className="flex md:absolute md:left-0 md:inset-y-0 lg:static xl:col-span-2">
-                    <div className="flex-shrink-0 flex items-center">
-                      <a href="#">
-                        <img
-                          className="block h-8 w-auto"
-                          src="https://tailwindui.com/img/logos/workflow-mark.svg?color=rose&shade=500"
-                          alt="Workflow"
-                        />
-                      </a>
-                    </div>
-                  </div>
-                  <div className="min-w-0 flex-1 md:px-8 lg:px-0 xl:col-span-6">
-                    <div className="flex items-center px-6 py-4 md:max-w-3xl md:mx-auto lg:max-w-none lg:mx-0 xl:px-0">
-                      <div className="w-full">
-                        <label htmlFor="search" className="sr-only">
-                          Search
-                        </label>
-                        <div className="relative">
-                          <div className="pointer-events-none absolute inset-y-0 left-0 pl-3 flex items-center">
-                            <SearchIcon
-                              className="h-5 w-5 text-gray-400"
-                              aria-hidden="true"
-                            />
-                          </div>
-                          <input
-                            id="search"
-                            name="search"
-                            className="block w-full bg-white border border-gray-300 rounded-md py-2 pl-10 pr-3 text-sm placeholder-gray-500 focus:outline-none focus:text-gray-900 focus:placeholder-gray-400 focus:ring-1 focus:ring-rose-500 focus:border-rose-500 sm:text-sm"
-                            placeholder="Search"
-                            type="search"
-                          />
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                  <div className="flex items-center md:absolute md:right-0 md:inset-y-0 lg:hidden">
-                    {/* Mobile menu button */}
-                    <Popover.Button className="-mx-2 rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:bg-gray-100 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-rose-500">
-                      <span className="sr-only">Open menu</span>
-                      {open ? (
-                        <XIcon className="block h-6 w-6" aria-hidden="true" />
-                      ) : (
-                        <MenuIcon
-                          className="block h-6 w-6"
-                          aria-hidden="true"
-                        />
-                      )}
-                    </Popover.Button>
-                  </div>
-                  <div className="hidden lg:flex lg:items-center lg:justify-end xl:col-span-4">
-                    <a
-                      href="#"
-                      className="text-sm font-medium text-gray-900 hover:underline"
-                    >
-                      Go Premium
-                    </a>
-                    <a
-                      href="#"
-                      className="ml-5 flex-shrink-0 bg-white rounded-full p-1 text-gray-400 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-rose-500"
-                    >
-                      <span className="sr-only">View notifications</span>
-                      <BellIcon className="h-6 w-6" aria-hidden="true" />
-                    </a>
-
-                    {/* Profile dropdown */}
-                    <Menu as="div" className="flex-shrink-0 relative ml-5">
-                      <div>
-                        <Menu.Button className="bg-white rounded-full flex focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-rose-500">
-                          <span className="sr-only">Open user menu</span>
-                          <img
-                            className="h-8 w-8 rounded-full"
-                            src={user.imageUrl}
-                            alt=""
-                          />
-                        </Menu.Button>
-                      </div>
-                      <Transition
-                        as={Fragment}
-                        enter="transition ease-out duration-100"
-                        enterFrom="transform opacity-0 scale-95"
-                        enterTo="transform opacity-100 scale-100"
-                        leave="transition ease-in duration-75"
-                        leaveFrom="transform opacity-100 scale-100"
-                        leaveTo="transform opacity-0 scale-95"
-                      >
-                        <Menu.Items className="origin-top-right absolute z-10 right-0 mt-2 w-48 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 py-1 focus:outline-none">
-                          {userNavigation.map((item) => (
-                            <Menu.Item key={item.name}>
-                              {({ active }) => (
-                                <a
-                                  href={item.href}
-                                  className={classnames(
-                                    active ? "bg-gray-100" : "",
-                                    "block py-2 px-4 text-sm text-gray-700"
-                                  )}
-                                >
-                                  {item.name}
-                                </a>
-                              )}
-                            </Menu.Item>
-                          ))}
-                        </Menu.Items>
-                      </Transition>
-                    </Menu>
-
-                    <a
-                      href="#"
-                      className="ml-6 inline-flex items-center px-4 py-2 border border-transparent text-sm font-medium rounded-md shadow-sm text-white bg-rose-600 hover:bg-rose-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-rose-500"
-                    >
-                      New Post
-                    </a>
-                  </div>
-                </div>
-              </div>
-
-              <Popover.Panel as="nav" className="lg:hidden" aria-label="Global">
-                <div className="max-w-3xl mx-auto px-2 pt-2 pb-3 space-y-1 sm:px-4">
-                  {navigation.map((item) => (
-                    <a
-                      key={item.name}
-                      href={item.href}
-                      aria-current={item.current ? "page" : undefined}
-                      className={classnames(
-                        item.current
-                          ? "bg-gray-100 text-gray-900"
-                          : "hover:bg-gray-50",
-                        "block rounded-md py-2 px-3 text-base font-medium"
-                      )}
-                    >
-                      {item.name}
-                    </a>
-                  ))}
-                </div>
-                <div className="border-t border-gray-200 pt-4">
-                  <div className="max-w-3xl mx-auto px-4 flex items-center sm:px-6">
-                    <div className="flex-shrink-0">
-                      <img
-                        className="h-10 w-10 rounded-full"
-                        src={user.imageUrl}
-                        alt=""
-                      />
-                    </div>
-                    <div className="ml-3">
-                      <div className="text-base font-medium text-gray-800">
-                        {user.name}
-                      </div>
-                      <div className="text-sm font-medium text-gray-500">
-                        {user.email}
-                      </div>
-                    </div>
-                    <button
-                      type="button"
-                      className="ml-auto flex-shrink-0 bg-white rounded-full p-1 text-gray-400 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-rose-500"
-                    >
-                      <span className="sr-only">View notifications</span>
-                      <BellIcon className="h-6 w-6" aria-hidden="true" />
-                    </button>
-                  </div>
-                  <div className="mt-3 max-w-3xl mx-auto px-2 space-y-1 sm:px-4">
-                    {userNavigation.map((item) => (
-                      <a
-                        key={item.name}
-                        href={item.href}
-                        className="block rounded-md py-2 px-3 text-base font-medium text-gray-500 hover:bg-gray-50 hover:text-gray-900"
-                      >
-                        {item.name}
-                      </a>
-                    ))}
-                  </div>
-                </div>
-
-                <div className="mt-6 max-w-3xl mx-auto px-4 sm:px-6">
-                  <a
-                    href="#"
-                    className="w-full flex items-center justify-center px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-rose-600 hover:bg-rose-700"
-                  >
-                    New Post
-                  </a>
-
-                  <div className="mt-6 flex justify-center">
-                    <a
-                      href="#"
-                      className="text-base font-medium text-gray-900 hover:underline"
-                    >
-                      Go Premium
-                    </a>
-                  </div>
-                </div>
-              </Popover.Panel>
-            </>
-          )}
-        </Popover>
-
-        <div className="py-10">
-          <div className="max-w-3xl mx-auto sm:px-6 lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-12 lg:gap-8">
-            <div className="hidden lg:block lg:col-span-3 xl:col-span-2">
-              <nav
-                aria-label="Sidebar"
-                className="sticky top-4 divide-y divide-gray-300"
-              >
-                <div className="pb-8 space-y-1">
-                  {navigation.map((item) => (
-                    <a
-                      key={item.name}
-                      href={item.href}
-                      className={classnames(
-                        item.current
-                          ? "bg-gray-200 text-gray-900"
-                          : "text-gray-600 hover:bg-gray-50",
-                        "group flex items-center px-3 py-2 text-sm font-medium rounded-md"
-                      )}
-                      aria-current={item.current ? "page" : undefined}
-                    >
-                      <item.icon
-                        className={classnames(
-                          item.current
-                            ? "text-gray-500"
-                            : "text-gray-400 group-hover:text-gray-500",
-                          "flex-shrink-0 -ml-1 mr-3 h-6 w-6"
-                        )}
-                        aria-hidden="true"
-                      />
-                      <span className="truncate">{item.name}</span>
-                    </a>
-                  ))}
-                </div>
-                <div className="pt-10">
-                  <p
-                    className="px-3 text-xs font-semibold text-gray-500 uppercase tracking-wider"
-                    id="communities-headline"
-                  >
-                    My communities
-                  </p>
-                  <div
-                    className="mt-3 space-y-2"
-                    aria-labelledby="communities-headline"
-                  >
-                    {communities.map((community) => (
-                      <a
-                        key={community.name}
-                        href={community.href}
-                        className="group flex items-center px-3 py-2 text-sm font-medium text-gray-600 rounded-md hover:text-gray-900 hover:bg-gray-50"
-                      >
-                        <span className="truncate">{community.name}</span>
-                      </a>
-                    ))}
-                  </div>
-                </div>
-              </nav>
-            </div>
-            <main className="lg:col-span-9 xl:col-span-6">
-              <div className="px-4 sm:px-0 mb-4">
-                <div className="sm:hidden">
-                  <label htmlFor="question-tabs" className="sr-only">
-                    Select a tab
-                  </label>
-                  <select
-                    id="question-tabs"
-                    className="block w-full rounded-md border-gray-300 text-base font-medium text-gray-900 shadow-sm focus:border-rose-500 focus:ring-rose-500"
-                    onChange={({ target: { value } }) => setTab(value)}
-                    value={tab}
-                  >
-                    {tabs.map((tab) => (
-                      <option key={tab.href} value={tab.href}>
-                        {tab.name}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-                <div className="hidden sm:block">
-                  <nav
-                    className="relative z-0 rounded-lg shadow flex divide-x divide-gray-200"
-                    aria-label="Tabs"
-                  >
-                    {tabs.map(({ name, href }, tabIdx) => (
-                      <a
-                        key={name}
-                        href={href}
-                        aria-current={href === tab ? "page" : undefined}
-                        className={classnames(
-                          href === tab
-                            ? "text-gray-900"
-                            : "text-gray-500 hover:text-gray-700",
-                          tabIdx === 0 ? "rounded-l-lg" : "",
-                          tabIdx === tabs.length - 1 ? "rounded-r-lg" : "",
-                          "group relative min-w-0 flex-1 overflow-hidden bg-white py-4 px-6 text-sm font-medium text-center hover:bg-gray-50 focus:z-10"
-                        )}
-                        onClick={() => setTab(href)}
-                      >
-                        <span>{name}</span>
-                        <span
-                          aria-hidden="true"
-                          className={classnames(
-                            href === tab ? "bg-rose-500" : "bg-transparent",
-                            "absolute inset-x-0 bottom-0 h-0.5"
-                          )}
-                        />
-                      </a>
-                    ))}
-                  </nav>
-                </div>
-              </div>
-              <TabPanel hidden={tab !== "#about"}>
-                <AboutMe />
-                <TechnicalSkills />
-              </TabPanel>
-              <TabPanel hidden={tab !== "#projects"}>
-                <Projects />
-              </TabPanel>
-              <TabPanel hidden={tab !== "#experience"}>
-                <Experience />
-              </TabPanel>
-            </main>
-            <aside className="hidden xl:block xl:col-span-4">
-              <div className="sticky top-4 space-y-4">
-                <section aria-labelledby="timeline-heading">
-                  <div className="bg-white rounded-lg shadow">
-                    <div className="p-6">
-                      <h2
-                        id="timeline-heading"
-                        className="text-base font-medium text-gray-900"
-                      >
-                        Timeline
-                      </h2>
-                      <div className="mt-6 flow-root">
-                        <Timeline />
-                      </div>
-                    </div>
-                  </div>
-                </section>
-              </div>
-            </aside>
-          </div>
-        </div>
-      </div>
-    </>
-  );
-}
